### 2018-05-09 #730 @whedon
[PRE REVIEW]: PyNomaly
**Submitting author:** @vc1492a (<a href="http://orcid.org/0000-0002-5279-4143">Valentino Constantinou</a>)
**Repository:** <a href="https://www.github.com/vc1492a/PyNomaly" target ="_blank">https://www.github.com/vc1492a/PyNomaly</a>
**Version:** v0.2.0
