### 2018-05-14 #730 @vc1492a
@arfon I was under the impression that references should be included in ```paper.bib```. The examples [linked from the JOSS website](https://github.com/arfon/fidgit/tree/master/paper) show references in ```paper.bib``` and not in ```paper.md```. Should I also add my references to ```paper.md```?
