### 2018-05-14 #730 @arfon
@vc1492a - I think the issue is that you're not citing them directly (so they're not showing up in the compiled PDF). You can read how to do that [here](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html)
