+++ paper/paper.bib
@@ -1,4 +1,4 @@
-@inproceedings{Breunig:2000:LID:342009.335388,
+@inproceedings{Breunig,
  author = {Breunig, Markus M. and Kriegel, Hans-Peter and Ng, Raymond T. and Sander, J\"{o}rg},
  title = {LOF: Identifying Density-based Local Outliers},
  booktitle = {Proceedings of the 2000 ACM SIGMOD International Conference on Management of Data},
@@ -16,7 +16,7 @@
  keywords = {database mining, outlier detection},
 }
 
-@inproceedings{Kriegel:2009:LLO:1645953.1646195,
+@inproceedings{Kriegel,
  author = {Kriegel, Hans-Peter and Kr\"{o}ger, Peer and Schubert, Erich and Zimek, Arthur},
  title = {LoOP: Local Outlier Probabilities},
  booktitle = {Proceedings of the 18th ACM Conference on Information and Knowledge Management},
@@ -34,23 +34,8 @@
  keywords = {outlier detection},
 }
 
-@article{Goldstein2016,
- doi = {10.1371/journal.pone.0152173},
- url = {https://doi.org/10.1371/journal.pone.0152173},
- year  = {2016},
- month = {apr},
- publisher = {Public Library of Science ({PLoS})},
- volume = {11},
- number = {4},
- pages = {e0152173},
- author = {Markus Goldstein and Seiichi Uchida},
- editor = {Dongxiao Zhu},
- title = {A Comparative Evaluation of Unsupervised Anomaly Detection Algorithms for Multivariate Data},
- journal = {{PLOS} {ONE}}
-}
-
-@article{
- doi:10.1080/23742917.2016.1226651,
+@article{Hamlet,
+ doi= {10.1080/23742917.2016.1226651},
  author = {Connor Hamlet and Jeremy Straub and Matthew Russell and Scott Kerlin},
  title = {An incremental and approximate local outlier probability algorithm for intrusion detection and its evaluation},
  journal = {Journal of Cyber Security Technology},
