+++ readme.md
@@ -1,8 +1,8 @@
 # PyNomaly
 
 PyNomaly is a Python 3 implementation of LoOP (Local Outlier Probabilities).
-LoOP is a local density based outlier detection method by Kriegel, Kröger, Schubert, and Zimek which provides outlier 
-scores in the range of [0,1] that are directly interpretable as the probability of a sample being an outlier. 
+LoOP is a local density based outlier detection method by Kriegel, Kröger, Schubert, and Zimek which provides outlier
+scores in the range of [0,1] that are directly interpretable as the probability of a sample being an outlier.
 
 [![PyPi](https://img.shields.io/badge/pypi-0.2.0-green.svg)](https://pypi.python.org/pypi/PyNomaly/0.2.0)
 [![License](https://img.shields.io/badge/License-Apache%202.0-blue.svg)](https://opensource.org/licenses/Apache-2.0)
@@ -11,25 +11,25 @@ The outlier score of each sample is called the Local Outlier Probability.
 It measures the local deviation of density of a given sample with
 respect to its neighbors as Local Outlier Factor (LOF), but provides normalized
 outlier scores in the range [0,1]. These outlier scores are directly interpretable
-as a probability of an object being an outlier. Since Local Outlier Probabilities provides scores in the 
+as a probability of an object being an outlier. Since Local Outlier Probabilities provides scores in the
 range [0,1], practitioners are free to interpret the results according to the application.
 
 Like LOF, it is local in that the anomaly score depends on how isolated the sample is
 with respect to the surrounding neighborhood. Locality is given by k-nearest neighbors,
-whose distance is used to estimate the local density. By comparing the local density of a sample to the 
+whose distance is used to estimate the local density. By comparing the local density of a sample to the
 local densities of its neighbors, one can identify samples that lie in regions of lower
-density compared to their neighbors and thus identify samples that may be outliers according to their Local 
+density compared to their neighbors and thus identify samples that may be outliers according to their Local
 Outlier Probability.
 
-The authors' 2009 paper detailing LoOP's theory, formulation, and application is provided by 
-Ludwig-Maximilians University Munich - Institute for Informatics; 
+The authors' 2009 paper detailing LoOP's theory, formulation, and application is provided by
+Ludwig-Maximilians University Munich - Institute for Informatics;
 [LoOP: Local Outlier Probabilities](http://www.dbs.ifi.lmu.de/Publikationen/Papers/LoOP1649.pdf).
 
 ## Implementation
 
-This Python 3 implementation uses Numpy and the formulas outlined in 
+This Python 3 implementation uses Numpy and the formulas outlined in
 [LoOP: Local Outlier Probabilities](http://www.dbs.ifi.lmu.de/Publikationen/Papers/LoOP1649.pdf)
-to calculate the Local Outlier Probability of each sample. 
+to calculate the Local Outlier Probability of each sample.
 
 ## Prerequisites
 - Python 3.5.2 or greater
@@ -45,7 +45,7 @@ First install the package from the Python Package Index:
 ```shell
 pip install PyNomaly # or pip3 install ... if you're using both Python 3 and 2.
 ```
-Then you can do something like this: 
+Then you can do something like this:
 
 ```python
 from PyNomaly import loop
@@ -53,7 +53,7 @@ m = loop.LocalOutlierProbability(data).fit()
 scores = m.local_outlier_probabilities
 print(scores)
 ```
-where *data* is a NxM (N rows, M columns) set of data as either a Pandas DataFrame or Numpy array. 
+where *data* is a NxM (N rows, M columns) set of data as either a Pandas DataFrame or Numpy array.
 
[…]
