### 2018-05-15 #730 @vc1492a
@afron thanks for the helpful link. I've updated my citations in ```paper.md```, let me know if it looks good to you.
