### 2018-05-19 #730 @arfon
@mlgill @sssomani - would you be willing to review this submission for JOSS?
