### 2018-05-20 #730 @arfon
@nhejazi @fritzo - would you be willing to review this submission for JOSS?
