### 2018-05-21 2fbf42d
Performance enhancements and error handling
-Reworked the _euclidean method to yield an ~2x performance increase
-Added some checks to circumvent an error that ocured when all input points are identical
