### 2018-05-26 cdb81ab
numpy fix
