### 2018-06-25 #730 @arfon
@jkahn - would you be willing to review this submission for JOSS?
