### 2018-06-25 #730 @Kevin-Mattheus-Moerman
@yzhao062 @betatim @aqibsaeed @rhiever would you be willing to review this submission for [JOSS](https://joss.theoj.org/about)? The [review process](https://joss.theoj.org/about#reviewer_guidelines) focuses largely on the software and a short paper.

[--> The paper :page_facing_up: <--](https://github.com/openjournals/joss-papers/blob/joss.00730/joss.00730/10.21105.joss.00730.pdf)

[--> The project repo :floppy_disk: : <--](https://github.com/vc1492a/PyNomaly)
