### 2018-06-25 #730 @yzhao062
@Kevin-Mattheus-Moerman I am happy to review PyNomaly since I used it in one of my previous research. Just need to get familiar with the [review process](https://joss.theoj.org/about#reviewer_guidelines) :)
