### 2018-06-25 #730 @Kevin-Mattheus-Moerman
@yzhao062 great thanks. I've added you as reviewer and will attempt to recruit a few more. Let me know if you have questions about the review process. This issue here is a so called pre-review issue. Once we have sufficient reviewers we'll move on to a review issue where the actual review takes place. I'll ping you there when we are ready.
The review process is guided by a set of tickboxes for each reviewer. Here is a nice example of a finished review process: https://github.com/openjournals/joss-reviews/issues/726
As you can see the reviewers comment in the review issue to provide feedback but also create PR's or issues on the project repository as well.
