### 2018-07-24 #845 @whedon
[REVIEW]: PyNomaly
**Submitting author:** @vc1492a (<a href="http://orcid.org/0000-0002-5279-4143">Valentino Constantinou</a>)
**Repository:** <a href="https://www.github.com/vc1492a/PyNomaly" target ="_blank">https://www.github.com/vc1492a/PyNomaly</a>
**Version:** v0.2.4

@yzhao062, @llllllllll, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:

---
