### 2018-07-25 #845 @vc1492a
@yzhao062 Current GitHub release is 0.2.1, as PyNomaly received an update after submission.
