### 2018-07-25 #845 @yzhao062
@Kevin-Mattheus-Moerman @vc1492a I am about to start by following the review schedule  above. Given I need some time to familiarize myself with the process, I would say 7-10 days is a safe estimate (should be faster though).
