+++ readme.md
@@ -6,6 +6,7 @@ scores in the range of [0,1] that are directly interpretable as the probability 
 
 [![PyPi](https://img.shields.io/badge/pypi-0.2.1-green.svg)](https://pypi.python.org/pypi/PyNomaly/0.2.1)
 [![License](https://img.shields.io/badge/License-Apache%202.0-blue.svg)](https://opensource.org/licenses/Apache-2.0)
+[![JOSS status](http://joss.theoj.org/papers/f4d2cfe680768526da7c1f6a2c103266/status.svg)](http://joss.theoj.org/papers/f4d2cfe680768526da7c1f6a2c103266)
 
 The outlier score of each sample is called the Local Outlier Probability.
 It measures the local deviation of density of a given sample with
