### 2018-08-01 #845 @yzhao062
Hi @Kevin-Mattheus-Moerman @vc1492a ,

I have followed the tasklist to conduct the first round review and marked most of the items as done. There are 2 items unchecked and want to list them below:

1. **Functionality documentation**: the parameters are introduced in readme.MD. Given there are only 2 parameters to tweak (extent and n_neighbors), it should be fine without a full documentation (e.g. readthedocs). However, it may be helpful to open another section to list the the default values (0.997 and 10) and your recommendations for picking the parameters.

2. **Automated tests**:  it could be helpful to include some basic tests to make sure the model is running. For example, it is easy to generate some pseudo data and use loOP to do a classification and verify the result, which could at least make sure it is doing the job on the simplest data. Feel free to use the [data_generation](https://github.com/yzhao062/Pyod/blob/master/pyod/utils/data.py) model here if that saves some time.

Other than those two items, I think it is a very solid and useful package to go which I used in my last paper. Since I am a first-time reviewer, please advise what is the best way to address the above two tasks.

Thanks,
Yue
