### 2018-08-17 #845 @vc1492a
@Kevin-Mattheus-Moerman Thanks for the message. I was under the impression that the work would be reviewed by a second reviewer prior to moving to the revision process. As such, I have not yet started on the recommendations @yzhao062 provided above but please let me know whether I should move forward with the revision process. I can get started in the next few days, and should be able to provide solutions to the above recommendations in about two weeks time.
