### 2018-08-30 e02c035
new examples, tests, buf fixes
