+++ PyNomaly/loop.py
@@ -4,16 +4,15 @@ import numpy as np
 import sys
 import warnings
 
-
 __author__ = 'Valentino Constantinou'
-__version__ = '0.2.1'
+__version__ = '0.2.2'
 __license__ = 'Apache License, Version 2.0'
 
 
 class LocalOutlierProbability(object):
     """
     :param data: a Pandas DataFrame or Numpy array of float data
-    :param extent: a parameter value in the range (0, 1] that controls the statistical extent (optional, default 0.997)
+    :param extent: an integer value [1, 2, 3] that controls the statistical extent, e.g. lambda times the standard deviation from the mean (optional, default 3)
     :param n_neighbors: the total number of neighbors to consider w.r.t. each sample (optional, default 10)
     :param cluster_labels: a numpy array of cluster assignments w.r.t. each sample (optional, default None)
     :return:
@@ -34,6 +33,88 @@ class LocalOutlierProbability(object):
            algorithm for intrusion detection and its evaluation. Journal of Cyber Security Technology (2016). 
     """
 
+    class Validate:
+
+        @staticmethod
+        def data(obj):
+            if obj.__class__.__name__ == 'DataFrame':
+                points_vector = obj.values
+                return points_vector
+            elif obj.__class__.__name__ == 'ndarray':
+                points_vector = obj
+                return points_vector
+            else:
+                warnings.warn(
+                    'Provided data must be in ndarray or DataFrame.',
+                    UserWarning)
+                if isinstance(obj, list):
+                    points_vector = np.array(obj)
+                    return points_vector
+                points_vector = np.array([obj])
+                return points_vector
+
+        @staticmethod
+        def n_neighbors(obj, set_neighbors=False):
+            if not obj.n_neighbors > 0:
+                warnings.warn('n_neighbors must be greater than 0.',
+                              UserWarning)
+                return False
+            elif obj.n_neighbors >= obj._n_observations():
+                if set_neighbors is True:
+                    obj.n_neighbors = obj._n_observations() - 1
+                warnings.warn(
+                    'n_neighbors must be less than the number of observations. Fit with ' + str(
+                        obj.n_neighbors) + ' instead.', UserWarning)
+                return True
+            for cluster_id in obj._cluster_labels():
+                c_labels = obj._cluster_labels()
+                if c_labels[np.where(np.array(c_labels) == cluster_id)].shape[
+                    0] <= obj.n_neighbors:
+                    warnings.warn(
+                        'Number of neighbors specified larger than smallest cluster. Specify a number of neighbors smaller than the smallest cluster size (observations in smallest cluster minus one).',
+                        UserWarning)
+                return False
+
+        @staticmethod
+        def extent(obj):
+            if obj.extent not in [1, 2, 3]:
+                warnings.warn(
+                    'extent parameter (lambda) must be 1, 2, or 3.',
+                    UserWarning)
+                return False
+            else:
+                return True
+
+        @staticmethod
+        def missing_values(obj):
+            if np.any(np.isnan(obj.data)):
+                warnings.warn(
+                    'Input data contains missing values. Some scores may not be returned.',
+                    UserWarning)
+                return False
+            else:
+                return True
+
+        @staticmethod
+        def fit(obj):
+            if obj.points_vector is None:
+                warnings.warn(
+                    "Must fit on historical data by calling fit() prior to calling stream(x).",
+                    UserWarning)
+                return False
+            else:
[…]
