+++ changelog.md
@@ -4,6 +4,26 @@ All notable changes to PyNomaly will be documented in this Changelog.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) 
 and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.2](#)
+### Fixed
+- [Issue #14](https://github.com/vc1492a/PyNomaly/issues/14) - Fixed an issue
+that was causing a ZeroDivisionError when the specified neighborhood size
+is larger than the total number of observations in the smallest cluster.
+
+### Changed
+- This implementation to align more closely with the specification of the
+approach in the original paper. The extent parameter now takes an integer
+value of 1, 2, or 3 that corresponds to the lambda parameter specified
+in the paper. See the [readme](https://github.com/vc1492a/PyNomaly/blob/master/readme.md) for more details.
+- Refactored the code base and created the Validate class, which includes
+checks for data type, correct specification, and other dependencies.
+
+### Added
+- Automated tests to ensure the desired functionality is being met can now be
+found in the `PyNomaly/tests` directory.
+- Code for the examples in the readme can now be found in the `examples` directory.
+- Additional information for parameter selection in the [readme](https://github.com/vc1492a/PyNomaly/blob/master/readme.md).
+
 ## [0.2.1](https://github.com/vc1492a/PyNomaly/commit/bb62253b94b4a3683cf64455905092a6a29a2088)
 ### Fixed
 - [Issue #10](https://github.com/vc1492a/PyNomaly/issues/10) - Fixed error on line

[…]
