### 2018-08-30 #845 @vc1492a
@Kevin-Mattheus-Moerman I have addressed the issues raised by @yzhao062. Additional documentation for parameter has been added to the readme, and a series of automated tests are available in the `tests` directory (using `pytest`). Additionally, examples have been added to the `examples` directory. Please let me know if there's any additional items I should address.
