### 2018-08-31 #845 @Kevin-Mattheus-Moerman
@yzhao062 could you respond to how @vc1492a has dealt with your feedback? Thanks
