+++ examples/multiple_gaussian_2d.py
@@ -28,15 +28,3 @@ for i in range(len(scores)):
 
 plt.show()
 
-
-'''
-The above plot shows that the results from this implementation generally 
-match in ranking vs the results shown in the original paper: the 
-probabilities are off by some margin in some cases. 
-Examination points to the calculation of the expected values during the 
-estimation process: where we calculate expected values using global means, 
-the paper mentions calculating the expected value over the neighborhood 
-about a particular observation. 
-'''
-
-
