+++ .travis.yml
@@ -1,13 +1,13 @@
 language: python
 python:
-  - "3.3"
   - "3.4"
   - "3.5"
   - "3.6"
-  - "3.7"
+  - "3.7-dev" # development branch
 # command to install dependencies
 install:
   - pip install -r requirements.txt
+  - python setup.py -q install
 # command to run tests
 script:
   - pytest
