+++ .travis.yml
@@ -3,11 +3,12 @@ python:
   - "3.4"
   - "3.5"
   - "3.6"
-  - "3.7-dev" # development branch
 # command to install dependencies
 install:
   - pip install -r requirements.txt
   - python setup.py -q install
+  - pip install pandas
+  - pip install scikit-learn
 # command to run tests
 script:
   - pytest
