### 2018-09-03 #845 @yzhao062
@Kevin-Mattheus-Moerman I have successfully run the newly added tests. I am confident it is a sound and useful package! Great work @vc1492a ! I think it is good to go:)
