+++ .coveralls.yml
@@ -1,1 +1,0 @@
-repo_token: gaY4IFes8Sk9tLlwR5DVWlNXX0yGnrwok
