### 2018-10-10 5fc7bed
DEV: use install_requires instead of setup_requires
