### 2018-10-11 #845 @llllllllll
Sorry it took me so long to get to this. This is a really cool project.

It looks like you just merged my install pr, so I will check that off. Thanks! I had one small comment about clarifying the layout of the input. It only took me a minute to figure it out once I played with it, but it made it hard to follow the examples.

For testing I used daily returns of US equities for the year of 2016. I played around with different ways of extracting features. I also tried using sector labels as my cluster labels. I was very happy to see that this library identified some known data quality issues, mostly around missing splits. More importantly, it found a few new issues. I will definitely try to use this more in the future. Thanks again, and sorry for taking so long on this.
