### 2018-10-11 #845 @vc1492a
@llllllllll Thanks for your comments and suggestions, and thanks for submitting the pr. I'll take a look at your opened issue regarding the layout of the input and will see what changes I can make to PyNomaly and/or the documentation to improve clarification.
