### 2018-10-14 7705232
input array shape clarification
