### 2018-10-14 #845 @vc1492a
@Kevin-Mattheus-Moerman I have added additional clarification in `readme.md` following @llllllllll's review and comments. Please let me know if there are any additional steps I should take, thanks!
