### 2018-10-26 #845 @Kevin-Mattheus-Moerman
This is looking good now. Thanks @yzhao062, @llllllllll for your review efforts! :rocket:

@vc1492a at this point could you please:
1) Inform us if the version number (v0.2.0) assigned here is still accurate or if you've since created a new release?
2) Deposit an archived version of the reviewed software in a service like [Zenodo](https://zenodo.org/) (see also: https://guides.github.com/activities/citable-code/) and report back with the DOI of the archived version.
