+++ setup.py
@@ -3,12 +3,12 @@ from setuptools import setup
 setup(
     name='PyNomaly',
     packages=['PyNomaly'],
-    version='0.2.3',
+    version='0.2.4',
     description='A Python 3 implementation of LoOP: Local Outlier Probabilities, a local density based outlier detection method providing an outlier score in the range of [0,1].',
     author='Valentino Constantinou',
     author_email='vc@valentino.io',
     url='https://github.com/vc1492a/PyNomaly',
-    download_url='https://github.com/vc1492a/PyNomaly/archive/0.2.3.tar.gz',
+    download_url='https://github.com/vc1492a/PyNomaly/archive/0.2.4.tar.gz',
     keywords=['outlier', 'anomaly', 'detection', 'machine', 'learning', 'probability'],
     classifiers=[],
     license='Apache License, Version 2.0',
