### 2018-10-27 #845 @vc1492a
@Kevin-Mattheus-Moerman Thanks for the update. The current version number is v0.2.4 - there's been a few releases since the start of the review process. I've deposited an archived version of the software, it's DOI is `10.5281/zenodo.1472519`. Besides the version number, the article proof looks good to me as well. Thanks @yzhao062 and @llllllllll for the review!
