### 2018-10-27 #845 @Kevin-Mattheus-Moerman
@arfon I recommend we accept this work in JOSS. Above is the archive DOI. Please also note the version number change. Thanks.
